Fix tests, compute Reynolds number as in Cigre207 report
