+++ linerate/models/cigre207.py
@@ -1,8 +1,10 @@
+from abc import abstractmethod
+
 from linerate import ThermalModel, Span, Weather
-from linerate.equations import solar_angles, cigre601, math, cigre207, dimensionless, convective_cooling
+from linerate.equations import solar_angles, cigre601, math, cigre207, dimensionless, convective_cooling, joule_heating
 from linerate.equations.math import switch_cos_sin
 from linerate.model import _copy_method_docstring
-from linerate.units import Date, Celsius, Ampere, WattPerMeter
+from linerate.units import Date, Celsius, Ampere, WattPerMeter, OhmPerMeter
 
 
 class Cigre207(ThermalModel):
@@ -63,8 +65,8 @@ class Cigre207(ThermalModel):
     ) -> WattPerMeter:
         D = self.span.conductor.conductor_diameter
         d = self.span.conductor.outer_layer_strand_diameter
-        y = self.span.conductor_altitude
         V = self.weather.wind_speed
+        y = self.span.conductor_altitude
         T_a = self.weather.air_temperature
         T_c = conductor_temperature
         T_f = 0.5 * (T_c + T_a)
@@ -80,7 +82,8 @@ class Cigre207(ThermalModel):
         # Reynolds number is defined in the text on page 5 of :cite:p:`cigre207`.
         # The definition includes a relative air density, which does not make sense, so we omit it here and use the
         # standard definition of Reynolds number instead.
-        Re = dimensionless.compute_reynolds_number(V, D, nu_f)
+        rho_r = cigre207.convective_cooling.compute_relative_air_density(y)
+        Re = cigre207.convective_cooling.compute_reynolds_number(V, D, nu_f, rho_r)
         Gr = dimensionless.compute_grashof_number(D, T_c, T_a, nu_f)
         Pr = cigre207.convective_cooling.compute_prandtl_number(T_f)
         Rs = dimensionless.compute_conductor_roughness(D, d)
@@ -114,4 +117,42 @@ class Cigre207(ThermalModel):
     ) -> WattPerMeter:
         return super().compute_radiative_cooling(
             conductor_temperature=conductor_temperature, current=current
+        )
+
+    @abstractmethod
+    def compute_resistance(self, conductor_temperature: Celsius, current: Ampere) -> OhmPerMeter:
+        r"""Compute the conductor resistance, :math:`R~\left[\Omega~\text{m}^{-1}\right]`.
+
+        Parameters
+        ----------
+        conductor_temperature:
+            :math:`T_\text{av}~\left[^\circ\text{C}\right]`. The average conductor temperature.
+        current:
+            :math:`I~\left[\text{A}\right]`. The current.
+
+        Returns
+        -------
+        Union[float, float64, ndarray[Any, dtype[float64]]]
+            :math:`R~\left[\Omega\right]`. The resistance at the given temperature and current.
+        """
+        resistance = joule_heating.compute_resistance(
+            conductor_temperature,
+            temperature1=self.span.conductor.temperature1,
+            temperature2=self.span.conductor.temperature2,
+            resistance_at_temperature1=self.span.conductor.resistance_at_temperature1,
+            resistance_at_temperature2=self.span.conductor.resistance_at_temperature2,
+        )
+
+        A = self.span.conductor.aluminium_cross_section_area
+        b = self.span.conductor.constant_magnetic_effect
+        m = self.span.conductor.current_density_proportional_magnetic_effect
+        max_increase = self.span.conductor.max_magnetic_core_relative_resistance_increase
+
+        return joule_heating.correct_resistance_acsr_magnetic_core_loss(
+            ac_resistance=resistance,
+            current=current,
+            aluminium_cross_section_area=A,
+            constant_magnetic_effect=b,
+            current_density_proportional_magnetic_effect=m,
+            max_relative_increase=max_increase,
         )
